+++ CHANGELOG.md
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.43.0] - 2025-10-06
 
 ### Added
-
+- Add GitHub webhook receivers for team-based alert routing to create GitHub issues for alerts with severity "ticket"
 - Add support for PrometheusOperator `ScrapeConfig` CRDs. This requires --stability.level=experimental
 - Add logging-enabled flag towards the logging-operator -> observability-operator merger.