+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changes
+
+- Alertmanager / PagerDuty: only send alerts that have to page
+
 ## [0.43.1] - 2025-10-07
 
 ### Fixed